--- voicevox_engine/morphing/MorphingException.py
@@ -1,4 +1,0 @@
-
-class MorphingException(Exception):
-    def __init__(self, message: str):
-        self.message = message
