--- voicevox_engine/morphing/__init__.py
@@ -1,15 +1,0 @@
-from .morphing import morphing
-from .create_morphing_pair_parameter import create_morphing_pair_parameter
-from .MorphingPairParameter import MorphingPairParameter
-from .MorphingQuery import MorphingQuery
-from .MorphingResult import MorphingResult
-from .MorphingException import MorphingException
-
-__all__ = [
-    'morphing',
-    'create_morphing_pair_parameter',
-    'MorphingPairParameter',
-    'MorphingQuery',
-    'MorphingResult',
-    'MorphingException',
-]
